Extract CsvStagingService + CandidatePersister; close Sprint 6 (#563 PR 6c)
Complete the RecruitmentCsvImporter modular decomposition. The staged
batched-import flow (ingest_job -> validate_job -> promote_batch ->
commit_job, plus the stale-job sweep) moves to a new
FreeFormCertificate\Recruitment\CsvStagingService, and the shared
candidate-persistence cluster (upsert_candidate, refresh_pcd_hash,
maybe_promote_candidate, build_adjutancy_map) moves to a new
FreeFormCertificate\Recruitment\CandidatePersister.

Both the synchronous run() flow and the staging promote_batch flow now
write candidates through CandidatePersister, eliminating the duplicated
lookup-then-create / encryption / wp_user-promotion logic.

RecruitmentCsvImporter becomes a thin facade: the four public staging
entry points (consumed by the REST controller + the batched test-suite)
and parse() delegate to the extracted classes, while run() /
import_preview() / import_definitive() / the failure*() envelopes and
BATCH_SIZE_DEFAULT stay as orchestration. No back-compat wrappers beyond
the public facade.

Behavior-preserving: full importer + batched characterization suites stay
green; full suite OK (5486 tests). The importer drops from ~1153 to ~310
lines (~1240 lines lifted into focused single-responsibility classes
across Sprint 6).
