+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format follows [Keep a Changelog] (https://keepachangelog.com/en/1.1.0/).
 - Internal refactor (#563 Sprint 5, PR 5a) — extracted the current-user permission gate from `Core\Utils` into a `FreeFormCertificate\Core\Capabilities` facade (B3): `current_user_can_manage()` and `current_user_can_admin_or()`. All ~100 call sites across ~48 files repointed (no back-compat wrappers); the `AjaxTrait` `class_exists` guards now reference `Capabilities`. Behavior-preserving thin facade over `current_user_can()`. (`CLAUDE.md` inline-gate guidance updated accordingly.)
 - Internal refactor (#563 Sprint 6, PR 6a) — began splitting the 1554-line `RecruitmentCsvImporter` by extracting its pure string→rows layer into a stateless `FreeFormCertificate\Recruitment\CsvParser`: `parse()`, `build_row()`, `parse_pcd_flag()`, `normalise_id()`, plus the `REQUIRED_HEADERS` / `OPTIONAL_HEADERS` contract (now public consts). The importer's public `parse()` delegates to the new class and every internal call site is repointed; none of these helpers touch the database or WordPress globals, making them the safest, most isolated slice to lift first. Behavior-preserving — the existing parse/normalise characterization tests stay green and a new `CsvParserTest` pins the previously-untested `parse_pcd_flag()` / `build_row()` leaves. The validator (PR 6b) and staging/commit services (PR 6c) follow.
 - Internal refactor (#563 Sprint 6, PR 6b) — extracted the §6 row-validation layer from `RecruitmentCsvImporter` into a stateless `FreeFormCertificate\Recruitment\CsvValidator`: the `validate()` rule set (CPF/RF presence + width, score / time-points format, rank, adjutancy membership, and the cross-row physical-candidate duplicate / field-divergence detection) plus the shared `line_error()` formatter (now a public static, since the staging ingest path uses it too). The importer's `run()` and staging paths delegate to `CsvValidator::`; the result-envelope builders (`failure()` / `failure_with_messages()`) stay on the importer as orchestrator-level concerns. Behavior-preserving — the existing validate characterization suite (repointed to the now-public method) stays green. The staging / commit services (PR 6c) follow.
+- Internal refactor (#563 Sprint 6, PR 6c) — completes the `RecruitmentCsvImporter` split (closes the Sprint 6 modular decomposition). The staged batched-import flow (`ingest_job` → `validate_job` → `promote_batch` → `commit_job`, plus the stale-job sweep) moves to `FreeFormCertificate\Recruitment\CsvStagingService`, and the shared candidate-persistence cluster (`upsert_candidate`, `refresh_pcd_hash`, `maybe_promote_candidate`, `build_adjutancy_map`) moves to `FreeFormCertificate\Recruitment\CandidatePersister` — both the synchronous `run()` flow and the staging `promote_batch` flow now write candidates through that one class, eliminating the duplicated lookup-then-create / encryption / wp_user-promotion logic. `RecruitmentCsvImporter` becomes a thin façade: the four public staging entry points (referenced by the REST controller + the batched test-suite) and `parse()` delegate to the extracted classes, while `run()` / `import_preview()` / `import_definitive()` / the `failure*()` envelopes and `BATCH_SIZE_DEFAULT` stay as orchestration. No back-compat wrappers beyond the public façade; behavior-preserving — the full importer + batched characterization suites stay green. The importer drops from ~1153 to ~310 lines (≈1240 lines lifted into focused single-responsibility classes across the sprint).
 
 ## [6.11.2] (2026-06-21)
